server/requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
Flask-SocketIO
Flask-SQLAlchemy
python-dotenv
eventlet
PyMySQL
